mask-detection/requirements.txt
Flask==1.1.2
google-python-cloud-debugger
google-cloud-storage==1.24.1
google-cloud-vision==0.33.0
pillow==7.1.2
opencv-python
werkzeug==1.0.1
flask-fs
connexion==2.6.0